requirements.txt
fastapi
uvicorn
python-multipart
sqlmodel
psycopg2-binary
python-dotenv
pillow
requests
openai